Add --no-gofmt option
Kunde21/markdownfmt attempts to reformat all snippets using the "go"
language type.

So it will rewrite anything like the following:

    ```go
    package main

    func foo() {
      // intentionally using two spaces
    }
    ```

This change adds a --no-gofmt command line option that opts out of this
behavior in markdownfmt by replacing the list of code formatters.

This change requires Kunde21/markdownfmt#38 to build.